refactor(external-video): extract duplicated startWatching code to service
