+++ bigbluebutton-html5/imports/ui/components/external-video-player/service.ts
@@ -1,18 +1,42 @@
 import ReactPlayer from 'react-player';
-
-import { Panopto } from './custom-players/panopto';
 import { ExternalVideo } from '/imports/ui/Types/meeting';
 
 const YOUTUBE_SHORTS_REGEX = new RegExp(/^(?:https?:\/\/)?(?:www\.)?(youtube\.com\/shorts)\/.+$/);
+const DAILYMOTION_MATCH_URL = /https?:\/\/(?:www\.)?dailymotion\.com\/video\/[a-zA-Z0-9]+(?:\?[^\s]*)?/g;
+const YOUTUBE_REGEX = new RegExp(/^(?:https?:\/\/)?(?:www\.)?(youtube\.com|youtu.be)\/.+$/);
+
+const startWatching = (startExternalVideoMutation) => {
+  return (url: string) => {
+    let externalVideoUrl = url;
+
+    if (YOUTUBE_SHORTS_REGEX.test(url)) {
+      const shortsUrl = url.replace('shorts/', 'watch?v=');
+      externalVideoUrl = shortsUrl;
+    }
+
+    if (YOUTUBE_REGEX.test(externalVideoUrl)) {
+      const YTUrl = new URL(externalVideoUrl);
+      YTUrl.searchParams.delete('list');
+      YTUrl.searchParams.delete('index');
+      externalVideoUrl = YTUrl.toString();
+    }
+
+    startExternalVideoMutation({ variables: { externalVideoUrl } });
+  }
+};
 
 const isUrlValid = (url: string) => {
   if (YOUTUBE_SHORTS_REGEX.test(url)) {
     const shortsUrl = url.replace('shorts/', 'watch?v=');
 
-    return /^https.*$/.test(shortsUrl) && (ReactPlayer.canPlay(shortsUrl) || Panopto.canPlay(shortsUrl));
+    return /^https.*$/.test(shortsUrl) && ReactPlayer.canPlay(shortsUrl);
   }
 
-  return /^https.*$/.test(url) && (ReactPlayer.canPlay(url) || Panopto.canPlay(url));
+  if (DAILYMOTION_MATCH_URL.test(url)) {
+    return false; // Dailymotion is not supported by react-player https://github.com/cookpete/react-player/issues/1772
+  }
+
+  return /^https.*$/.test(url) && ReactPlayer.canPlay(url);
 };
 
 // Convert state (Number) to playing (Boolean)
@@ -41,4 +65,5 @@ export {
   isUrlValid,
   getPlayingState,
   calculateCurrentTime,
+  startWatching,
 };
