+++ bigbluebutton-html5/imports/ui/components/presentation/presentation-toolbar/smart-video-share/container.jsx
@@ -1,30 +1,15 @@
 import React from 'react';
 import { useMutation } from '@apollo/client';
 import { SmartMediaShare } from './component';
-import Panopto from '../../../external-video-player/custom-players/panopto';
-import { EXTERNAL_VIDEO_START } from '../../../external-video-player/mutations';
-
-const YOUTUBE_SHORTS_REGEX = new RegExp(/^(?:https?:\/\/)?(?:www\.)?(youtube\.com\/shorts)\/.+$/);
+import { startWatching } from '../../../external-video-player/service';
+import { EXTERNAL_VIDEO_START } from '/imports/ui/components/external-video-player/mutations';
 
 const SmartMediaShareContainer = (props) => {
-  const [startExternalVideo] = useMutation(EXTERNAL_VIDEO_START);
-
-  const startWatching = (url) => {
-    let externalVideoUrl = url;
-
-    if (YOUTUBE_SHORTS_REGEX.test(url)) {
-      const shortsUrl = url.replace('shorts/', 'watch?v=');
-      externalVideoUrl = shortsUrl;
-    } else if (Panopto.canPlay(url)) {
-      externalVideoUrl = Panopto.getSocialUrl(url);
-    }
-
-    startExternalVideo({ variables: { externalVideoUrl } });
-  };
+  const [startExternalVideoMutation] = useMutation(EXTERNAL_VIDEO_START);
 
   return (
     <SmartMediaShare {...{
-      startWatching,
+      startWatching: startWatching(startExternalVideoMutation),
       ...props,
     }}
     />
